requirements.txt
fastapi
uvicorn[standard]
gradio
requests
transformers
torch
Pillow
python-multipart